update arch install system tools script

+++ Arch-Install/install.sh
@@ -5,41 +5,63 @@
 # Source: https://www.github.com/chrporter22/dotfiles
 # Author: Christian J. Porter
 
-
-# Main function to handle installation process
-main() {
-    start_installation
-    prompt_reboot
+# Launch installation
+kickstart() {
+    sys_prep
+    xorg_stack
+    wm_suite
+    post_tools
+    dotfile_link
+    reboot_prompt
 }
 
-
-# Function to prompt for reboot
-prompt_reboot() {
-    echo -e "Installation successful. Do you want to reboot? (Y/n)"
-    read reboot
-
-    if [[ $reboot == "Y" || $reboot == "y" ]]; then
-        reboot
-    fi
+# Prompt to reboot
+reboot_prompt() {
+    echo -e "Installation complete. Reboot now? (Y/n)"
+    read -r ans
+    [[ $ans =~ ^[Yy]$ ]] && reboot
 }
 
-
-# Function to install system packages
-install_sys_packages() {
-    sudo pacman -Syu --needed base-devel neovim vim github-cli fastfetch firefox git wget curl htop feh flatpak ueberzug atool poppler imagemagick highlight zathura-pdf-poppler pacman-contrib  bluez bluez-utils 
+# System essentials
+sys_prep() {
+    sudo pacman -Syu --needed base-devel git curl wget firefox \
+        neovim vim github-cli fastfetch htop feh flatpak ueberzug \
+        atool poppler imagemagick highlight zathura-pdf-poppler \
+        pacman-contrib bluez bluez-utils
 }
 
-
-# Function to install AUR helper
-install_yay() {
-    git clone https://aur.archlinux.org/yay.git
-    cd yay || exit
-    makepkg -si --noconfirm
-    cd ..
+# Xorg & display server components
+xorg_stack() {
+    sudo pacman -S --needed xorg-server xorg-xinit xclip xdotool maim \
+        libxcb arandr ttf-firacode-nerd
 }
 
-# Function to install Xorg (window server/manager)
-install_xorg() {
-    sudo pacman -Syu xorg maim xclip xdotool picom libxcb arandr 
-    yay -S networkmanager-dmenu-git xbanish ttf-firacode-nerd
+# Window manager, compositor, launcher, bar
+wm_suite() {
+    sudo pacman -S --needed \
+        i3-wm i3status \
+        rofi \
+        picom \       # Compositor for transparency and effects
+        polybar \     # Customizable status bar
+        xbanish       # Hides mouse cursor when typing
+}
+
+# Post-install system tools and terminal
+post_tools() {
+    # Audio
+    sudo pacman -S --needed pulseaudio pulseaudio-equalizer pulseaudio-jack \
+        pulseaudio-alsa alsa-utils
+
+    # Terminal + Font
+    sudo pacman -S --needed alacritty ttf-jetbrains-mono-nerd
+
+    # Verify font installation
+    fc-list | grep "JetBrainsMono"
+}
+
+# Symlink dotfiles using stow
+dotfile_link() {
+    sudo pacman -S --needed stow
+    cd ~/dotfiles || exit
+    stow i3 picom polybar rofi alacritty xresources
 }
